requirements.txt
# Add required dependencies here
tensorflow==2.12.0
numpy==1.23.5
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
pillow==9.5.0
flask==2.3.2
flask-cors==3.0.10
joblib==1.2.0
werkzeug==2.3.4